Vegas Data Set: 2007-2011 (Wins)
I've compiled a list of all the overtime-winning teams from 2007-2011 and their Vegas odds, which include the spread at kickoff, whether the home team covered the line, the over-under, and the result of the over-under for the winning team. 

Obvs all the winning teams covered the spread so I'll add ties in a sec (doing "any" in Pro Football Reference causes double counts so I wanted to make 2 separate data sets for winning and drawing).